+++ src/test/random_tests.cpp
@@ -57,7 +57,7 @@ BOOST_FIXTURE_TEST_SUITE(random_tests, BasicTestingSetup)
             for (int j = 0; j < 1000; ++j)
             {
                 uint64_t rangebits = ctx1.randbits(bits);
-                BOOST_CHECK_EQUAL(rangebits >> bits, 0);
+                BOOST_CHECK_EQUAL(rangebits >> bits, (uint64_t)0);
                 uint64_t range = ((uint64_t) 1) << bits | rangebits;
                 uint64_t rand = ctx2.randrange(range);
                 BOOST_CHECK(rand < range);
