+++ src/test/univalue_tests.cpp
@@ -120,13 +120,13 @@ BOOST_FIXTURE_TEST_SUITE(univalue_tests, BasicTestingSetup)
 
         BOOST_CHECK(v.setObject());
         BOOST_CHECK(v.isObject());
-        BOOST_CHECK_EQUAL(v.size(), 0);
+        BOOST_CHECK_EQUAL(v.size(), (uint64_t)0);
         BOOST_CHECK_EQUAL(v.getType(), UniValue::VOBJ);
         BOOST_CHECK(v.empty());
 
         BOOST_CHECK(v.setArray());
         BOOST_CHECK(v.isArray());
-        BOOST_CHECK_EQUAL(v.size(), 0);
+        BOOST_CHECK_EQUAL(v.size(), (uint64_t)0);
 
         BOOST_CHECK(v.setStr("zum"));
         BOOST_CHECK(v.isStr());
@@ -195,7 +195,7 @@ BOOST_FIXTURE_TEST_SUITE(univalue_tests, BasicTestingSetup)
         BOOST_CHECK(arr.push_backV(vec));
 
         BOOST_CHECK_EQUAL(arr.empty(), false);
-        BOOST_CHECK_EQUAL(arr.size(), 5);
+        BOOST_CHECK_EQUAL(arr.size(), (uint64_t)5);
 
         BOOST_CHECK_EQUAL(arr[0].getValStr(), "1023");
         BOOST_CHECK_EQUAL(arr[1].getValStr(), "zippy");
@@ -207,7 +207,7 @@ BOOST_FIXTURE_TEST_SUITE(univalue_tests, BasicTestingSetup)
 
         arr.clear();
         BOOST_CHECK(arr.empty());
-        BOOST_CHECK_EQUAL(arr.size(), 0);
+        BOOST_CHECK_EQUAL(arr.size(), (uint64_t)0);
     }
 
     BOOST_AUTO_TEST_CASE(univalue_object_test)
@@ -249,7 +249,7 @@ BOOST_FIXTURE_TEST_SUITE(univalue_tests, BasicTestingSetup)
         BOOST_CHECK(obj.pushKVs(obj2));
 
         BOOST_CHECK_EQUAL(obj.empty(), false);
-        BOOST_CHECK_EQUAL(obj.size(), 9);
+        BOOST_CHECK_EQUAL(obj.size(), (uint64_t)9);
 
         BOOST_CHECK_EQUAL(obj["age"].getValStr(), "100");
         BOOST_CHECK_EQUAL(obj["first"].getValStr(), "John");
@@ -292,7 +292,7 @@ BOOST_FIXTURE_TEST_SUITE(univalue_tests, BasicTestingSetup)
 
         obj.clear();
         BOOST_CHECK(obj.empty());
-        BOOST_CHECK_EQUAL(obj.size(), 0);
+        BOOST_CHECK_EQUAL(obj.size(), (uint64_t)0);
     }
 
     static const char *json1 =
@@ -309,13 +309,13 @@ BOOST_FIXTURE_TEST_SUITE(univalue_tests, BasicTestingSetup)
         BOOST_CHECK(v.read(strJson1));
 
         BOOST_CHECK(v.isArray());
-        BOOST_CHECK_EQUAL(v.size(), 2);
+        BOOST_CHECK_EQUAL(v.size(), (uint64_t)2);
 
         BOOST_CHECK_EQUAL(v[0].getValStr(), "1.10000000");
 
         UniValue obj = v[1];
         BOOST_CHECK(obj.isObject());
-        BOOST_CHECK_EQUAL(obj.size(), 3);
+        BOOST_CHECK_EQUAL(obj.size(), (uint64_t)3);
 
         BOOST_CHECK(obj["key1"].isStr());
         std::string correctValue("str");
